requirements.txt
cryptography
django
django-bootstrap5
django-crispy-forms
django-fsm
django-filter
djangorestframework
djangorestframework-simplejwt
fontawesomefree
markdown
mozilla-django-oidc
paramiko
psycopg2-binary
uwsgi
